Sorta fix teams pages
Due to react router 6 update. Some stuff is still not working, but this
will at least get most things to render. Need to revisit later.
