+++ apps/client/assets/js/components/TeamMembershipForm.jsx
@@ -68,7 +68,7 @@ class _TeamMembershipForm extends React.Component {
         <div>
           {Object.keys(this.props.teams).map((id) => (
             <div key={id}>
-              <Link to={`teams/${id}`}>{this.props.teams[id].name}</Link>
+              <Link to={`/teams/${id}`}>{this.props.teams[id].name}</Link>
             </div>
           ))}
         </div>
